requirements.txt
django>=4.2
channels[daphne]>=4.0.0
djangorestframework
django-cors-headers
groq
python-dotenv
daphne